+++ fsFirstExistingFile.py
@@ -1,7 +1,7 @@
-import FileSystem;
+import mFileSystem;
 def fsFirstExistingFile(*asPossiblePaths):
   for sPossiblePath in asPossiblePaths:
-    if sPossiblePath and FileSystem.fbIsFile(sPossiblePath):
+    if sPossiblePath and mFileSystem.fbIsFile(sPossiblePath):
       return sPossiblePath;
   return None;