+++ modules/mFileSystem/modules/mWindowsAPI/StructureTypes.py
@@ -1,0 +1,173 @@
+import ctypes;
+from Defines import *;
+from Functions import *;
+from PrimitiveTypes import *;
+
+uNamelessStructureOrUnionsCounter = 0;
+def fcStructureOrUnion(sStructure_or_Union, sStructureName, axFields):
+  global uNamelessStructureOrUnionsCounter;
+  if sStructureName is None:
+    uNamelessStructureOrUnionsCounter += 1;
+    sStructureName = "_nameless_%s_%d_" % (sStructure_or_Union.lower(), uNamelessStructureOrUnionsCounter);
+  asAnonymousFieldNames = [];
+  atxFields = [];
+  for xField in axFields:
+    if isinstance(xField, tuple):
+      cFieldType, sFieldName = xField;
+    else:
+      cFieldType = xField;
+      sFieldName = "_anonymous_field_%d_" % len(asAnonymousFieldNames);
+      asAnonymousFieldNames.append(sFieldName);
+    atxFields.append((sFieldName, cFieldType));
+  cStructure_or_Union = sStructure_or_Union == "Structure" and ctypes.Structure or ctypes.Union;
+  return type(
+    sStructureName,
+    (cStructure_or_Union,),
+    {
+      "_anonymous_": asAnonymousFieldNames,
+      "_fields_": atxFields,
+    },
+  );
+
+def UNION(*axFields):
+  return fcStructureOrUnion("Union", None, axFields);
+
+def STRUCT(*axFields):
+  return fcStructureOrUnion("Structure", None, axFields);
+
+# Defining a structure also defines "P<struct_name>" and "PP<struct_name>" as a
+# pointer-to-structure and a pointer-to-pointer-to-structure respectively.
+def fDefineStructure(sName, *atxFields):
+  cStructure = fcStructureOrUnion("Structure", sName, atxFields);
+  globals()[sName] = cStructure;
+  globals()["LP" + sName] = POINTER(cStructure);
+  globals()["P" + sName] = POINTER(cStructure);
+  globals()["PP" + sName] = POINTER(POINTER(cStructure));
+
+################################################################################
+# Simple structures that contain only primitives and no other structures are   #
+# defined first, strcutures that contain other structures must wait until      #
+# those structures are defined and will therefore be defined in a second round #
+################################################################################
+
+#CCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCC
+fDefineStructure("COORD", 
+  (SHORT,       "X"),
+  (SHORT,       "Y"),
+);
+
+#FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF
+fDefineStructure("FILETIME",
+  (DWORD,       "dwLowDateTime"),
+  (DWORD,       "dwHighDateTime"),
+);
+#MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
+fDefineStructure("MODULEENTRY32A",
+  (DWORD,       "dwSize"),
+  (DWORD,       "th32ModuleID"),
+  (DWORD,       "th32ProcessID"),
+  (DWORD,       "GlblcntUsage"),
+  (DWORD,       "ProccntUsage"),
+  (PBYTE,       "modBaseAddr"),
+  (DWORD,       "modBaseSize"),
+  (HMODULE,     "hModule"),
+  (CHAR * (MAX_MODULE_NAME32 + 1), "szModule"),
+  (CHAR * MAX_PATH, "szExePath"),
+);
+fDefineStructure("MODULEENTRY32W",
+  (DWORD,       "dwSize"),
+  (DWORD,       "th32ModuleID"),
+  (DWORD,       "th32ProcessID"),
+  (DWORD,       "GlblcntUsage"),
+  (DWORD,       "ProccntUsage"),
+  (PBYTE,       "modBaseAddr"),
+  (DWORD,       "modBaseSize"),
+  (HMODULE,     "hModule"),
+  (WCHAR * (MAX_MODULE_NAME32 + 1), "szModule"),
+  (WCHAR * MAX_PATH, "szExePath"),
+);
+#OOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOO
+fDefineStructure("OVERLAPPED",
+  (PULONG,      "Internal"),
+  (PULONG,      "InternalHigh"),
+  UNION(
+    STRUCT(
+      (DWORD, "Offset"),
+      (DWORD, "OffsetHigh"),
+    ),
+    (PVOID, "Pointer"),
+  ),
+  (HANDLE,      "hEvent"),
+);
+#PPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPPP
+fDefineStructure("POINT",
+  (LONG,        "x"),
+  (LONG,        "y"),
+);
+fDefineStructure("PROCESSENTRY32A",
+  (DWORD,       "dwSize"),
+  (DWORD,       "cntUsage"),
+  (DWORD,       "th32ProcessID"),
+  (PULONG,      "th32DefaultHeapID"),
+  (DWORD,       "th32ModuleID"),
+  (DWORD,       "cntThreads"),
+  (DWORD,       "th32ParentProcessID"),
+  (LONG,        "pcPriClassBase"),
+  (DWORD,       "dwFlags"),
+  (CHAR * MAX_PATH, "szExeFile"),
+);
+fDefineStructure("PROCESSENTRY32W",
+  (DWORD,       "dwSize"),
+  (DWORD,       "cntUsage"),
+  (DWORD,       "th32ProcessID"),
+  (PULONG,      "th32DefaultHeapID"),
+  (DWORD,       "th32ModuleID"),
+  (DWORD,       "cntThreads"),
+  (DWORD,       "th32ParentProcessID"),
+  (LONG,        "pcPriClassBase"),
+  (DWORD,       "dwFlags"),
+  (WCHAR * MAX_PATH, "szExeFile"),
+);
+#RRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRR
+fDefineStructure("RECT",
+  (LONG,        "left"),
+  (LONG,        "top"),
+  (LONG,        "right"),
+  (LONG,        "bottom"),
+);
+
+#SSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSS
+fDefineStructure("SID");
+fDefineStructure("SIZE",
+  (LONG,        "cx"),
+  (LONG,        "cy"),
+);
+fDefineStructure("SMALL_RECT",
+  (SHORT,       "Left"),
+  (SHORT,       "Top"),
+  (SHORT,       "Right"),
+  (SHORT,       "Bottom"),
+);
+
+################################################################################
+# Structures that contain other structures                                     #
+################################################################################
+
+#CCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCC
+fDefineStructure("CONSOLE_SCREEN_BUFFER_INFO", 
+  (COORD,       "dwSize"),
+  (COORD,       "dwCursorPosition"),
+  (WORD,        "wAttributes"),
+  (SMALL_RECT,  "srWindow"),
+  (COORD,       "dwMaximumWindowSize"),
+);
+#SSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSSS
+fDefineStructure("SID_AND_ATTRIBUTES",
+  (PSID,        "Sid"),
+  (DWORD,       "Attributes"),
+);
+#TTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTT
+fDefineStructure("TOKEN_MANDATORY_LABEL",
+  (SID_AND_ATTRIBUTES, "Label"),
+);
+