+++ modules/mFileSystem/modules/mWindowsAPI/__init__.py
@@ -1,0 +1,9 @@
+from Defines import *;
+from Functions import *;
+from PrimitiveTypes import *;
+from StructureTypes import *;
+
+from ADVAPI32 import ADVAPI32;
+from KERNEL32 import KERNEL32;
+
+from oVersionInformation import oVersionInformation;